images moved to assets
